Закончил логику, подкотовил фронт к показу

+++ backend/BetPrediction/BetPrediction/Models/Dto/ProPlayerDto.cs
@@ -8,7 +8,7 @@ public class ProPlayerDto
     public string TeamName { get; set; }
     public string Role { get; set; }
     public string Region { get; set; }
-    public int Prizes { get; set; }
+    public double Prizes { get; set; }
     public int MatchesCount { get; set; }
     public int WinsCount { get; set; }
     public int LossesCount { get; set; }
